+++ src/lib/apputils/Makefile.in
@@ -12,19 +12,20 @@ SED = sed
 ##DOS##XTRA=
 ##DOS##OBJFILE=$(OUTPRE)apputils.lst
 
-STLIBOBJS=net-server.o @LIBOBJS@
+STLIBOBJS=net-server.o udppktinfo.o @LIBOBJS@
 LIBBASE=apputils
 
 all-unix:: all-liblinks
 clean-unix:: clean-liblinks clean-libs clean-libobjs
 install-unix:: install-libs
 
-LINTFLAGS=-uhvb 
+LINTFLAGS=-uhvb
 LINTFILES= daemon.c
 LIBOBJS=$(OUTPRE)daemon.$(OBJEXT)
 
 SRCS=	$(srcdir)/daemon.c \
-	$(srcdir)/net-server.c
+	$(srcdir)/net-server.c \
+	$(srcdir)/udppktinfo.c
 
 @libpriv_frag@
 @lib_frag@
